requirements.txt
ultralytics==8.1.24
Pillow
pillow-avif-plugin
dill
torch==2.2.1
torchvision==0.17.1